+++ plonky2-backend/src/plonky2_ecdsa/biguint/gates/arithmetic_u32.rs
@@ -36,7 +36,7 @@ impl<F: RichField + Extendable<D>, const D: usize> U32ArithmeticGate<F, D> {
         }
     }
 
-    pub(crate) fn num_ops(config: &CircuitConfig) -> usize {
+    pub fn num_ops(config: &CircuitConfig) -> usize {
         let wires_per_op = Self::routed_wires_per_op() + Self::num_limbs();
         (config.num_wires / wires_per_op).min(config.num_routed_wires / Self::routed_wires_per_op())
     }
@@ -446,7 +446,7 @@ impl<F: RichField + Extendable<D>, const D: usize> SimpleGenerator<F, D>
 
 #[cfg(test)]
 mod tests {
-    use crate::biguint::gates::gate_testing::{test_eval_fns, test_low_degree};
+    use crate::plonky2_ecdsa::biguint::gates::gate_testing::{test_eval_fns, test_low_degree};
     use anyhow::Result;
     use plonky2::field::goldilocks_field::GoldilocksField;
     use plonky2::field::types::Sample;
